+++ EDDI/ChangeLog.md
@@ -6,6 +6,7 @@
     * Add SpeechPitch() Cottle function. Details on how to use this are in the SpeechResponder documentation
     * Add SpeechRate() Cottle function. Details on how to use this are in the SpeechResponder documentation
     * Add SpeechVolume() Cottle function. Details on how to use this are in the SpeechResponder documentation
+    * Fixed a bug that was causing EDDI to occasionally not use Cereproc voices correctly / revert to a different voice
     * Fixed a bug that was preventing the Play() function from working properly
 
 ### 2.4.3